Modernize repo
Brings repo up to date.

- Fixed test namespaces and imports.
- Added tools.
- Upgraded Docker env.
- Now building with GH Actions instead of Travis.
