--- .codeclimate.yml
@@ -1,17 +1,0 @@
----
-engines:
-  duplication:
-    enabled: true
-    config:
-      languages:
-      - php
-  fixme:
-    enabled: true
-  phpmd:
-    enabled: true
-ratings:
-  paths:
-  - "**.inc"
-  - "**.php"
-exclude_paths:
-- test/
