+++ .env.example
@@ -1,1 +1,8 @@
 BASE_PATH=./
+BUILD_ROOT_PATH=/app/
+PROJECT_NAME=dhii_output-renderer-interface
+
+PHP_BUILD_VERSION=7.2
+PHP_TEST_VERSION=7.2
+
+
