+++ .github/workflows/continuous-integration.yml
@@ -1,0 +1,34 @@
+name: Continuous Integration
+on: [push, pull_request]
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        php-versions: ['7.2', '7.3', '7.4', '8.0']
+
+    steps:
+    - uses: actions/checkout@v2
+
+    - name: Setup PHP
+      uses: shivammathur/setup-php@v2
+      with:
+        php-version: ${{ matrix.php-versions }}
+
+    - name: Analysing source code
+      run: find ./src/ ./inc/ ./tests/ -type f -name '*.php' -print0 | xargs -0 -L 1 -P 4 -- php -l
+
+    - name: Validate composer.json and composer.lock
+      run: composer validate
+
+    - name: Install dependencies
+      run: composer install --prefer-dist --no-progress --no-suggest
+
+    - name: PhpUnit
+      run: ./vendor/bin/phpunit
+
+    - name: Psalm
+      run: ./vendor/bin/psalm --show-info=false --threads=8 --diff --diff-methods
+
+    - name: PHPCS
+      run: ./vendor/bin/phpcs -s --report-source --runtime-set ignore_warnings_on_exit 1
