+++ .gitignore
@@ -1,9 +1,10 @@
 /vendor/
 /nbproject/private/
-/test/coverage/
-/test/log/
+/tests/coverage/
+/tests/log/
 /.idea/inspectionProfiles/
 /.idea/codeStyles/
 /.idea/workspace.xml
+/.phpunit.result.cache
 
 /.env
