--- .travis.yml
@@ -1,29 +1,0 @@
-dist: precise
-language: php
-php:
-  - '7.0'
-  - '7.1'
-  - '7.2'
-  - '7.3'
-  - '7.4snapshot'
-  - nightly
-
-matrix:
-  allow_failures:
-    - php: nightly
-    - php: '7.4snapshot'
-  fast_finish: true
-
-before_script:
-  - phpenv config-add travis.php.ini
-  - composer update --prefer-dist
-
-script:
-  - vendor/bin/phpunit
-
-after_script:
-  - vendor/bin/test-reporter --coverage-report="test/coverage/clover.xml"
-
-cache:
-  directories:
-    - $HOME/.composer/cache/files
