+++ composer.json
@@ -12,14 +12,14 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^7.0",
+        "php": "^7.0 | ^8.0",
         "dhii/stringable-interface": "^0.1"
     },
     "require-dev": {
-        "dhii/php-cs-fixer-config": "dev-php-5.3",
-        "phpunit/phpunit": "^6.0",
-        "codeclimate/php-test-reporter": "dev-master",
-        "psr/container": "^1.0"
+        "phpunit/phpunit": "^6.0 | ^7.0 | ^8.0 | ^9.0",
+        "psr/container": "^1.0",
+        "vimeo/psalm": "^3.11.7",
+        "slevomat/coding-standard": "^6.0"
     },
     "autoload": {
         "psr-4": {
@@ -28,9 +28,9 @@
     },
     "autoload-dev": {
         "psr-4": {
-            "Dhii\\Output\\UnitTest\\": "test/unit",
-            "Dhii\\Output\\FuncTest\\": "test/functional",
-            "Dhii\\Output\\Test\\": "test/stubs"
+            "Dhii\\Output\\UnitTest\\": "tests/unit",
+            "Dhii\\Output\\FuncTest\\": "tests/functional",
+            "Dhii\\Output\\Test\\": "tests/stubs"
         }
     },
     "extra": {
