+++ docker-compose.yml
@@ -1,8 +1,29 @@
-version: '3.1'
+version: '3.4'
 services:
-  php:
+
+  build:
     build:
-      context: docker
-      dockerfile: Dockerfile_php
+      context: ./
+      dockerfile: docker/Dockerfile
+      target: build
+      args:
+        PHP_BUILD_VERSION: $PHP_BUILD_VERSION
+        BUILD_ROOT_PATH: $BUILD_ROOT_PATH
+    container_name: "${PROJECT_NAME}_build"
+    working_dir: ${BUILD_ROOT_PATH}
     volumes:
-      - ${BASE_PATH}:/srv/
+      -  ${BASE_PATH}:${BUILD_ROOT_PATH}
+
+  test:
+    build:
+      context: ./
+      dockerfile: docker/Dockerfile
+      target: test
+      args:
+        BUILD_ROOT_PATH: $BUILD_ROOT_PATH
+        PHP_BUILD_VERSION: $PHP_BUILD_VERSION
+        PHP_TEST_VERSION: $PHP_TEST_VERSION
+    container_name: "${PROJECT_NAME}_test"
+    working_dir: ${BUILD_ROOT_PATH}
+    volumes:
+      -  ${BASE_PATH}:${BUILD_ROOT_PATH}
