+++ tests/stubs/GetImplementingMockBuilderCapableTrait.php
