+++ tests/unit/Block/BlockAwareInterfaceTest.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Dhii\Block\UnitTest\Template;
+namespace Dhii\Output\UnitTest\Block;
 
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Framework_MockObject_MockObject as MockObject;
-use Dhii\Output\Template\PathTemplateFactoryInterface as TestSubject;
+use PHPUnit\Framework\MockObject\MockObject as MockObject;
+use Dhii\Output\Block\BlockAwareInterface as TestSubject;
 
 /**
  * Tests {@see TestSubject}.
  *
  * @since 0.4
  */
-class PathTemplateFactoryInterfaceTest extends TestCase
+class BlockAwareInterfaceTest extends TestCase
 {
     /**
      * Creates a new instance of the test subject.
