+++ tests/unit/ContextAwareInterfaceTest.php
@@ -1,22 +1,22 @@
 <?php
 
-namespace Dhii\Block\UnitTest\Block;
+namespace Dhii\Output\UnitTest;
 
+use Dhii\Output\ContextAwareInterface as TestSubject;
+use PHPUnit\Framework\MockObject\MockObject as MockObject;
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Framework_MockObject_MockObject as MockObject;
-use Dhii\Output\Block\StringBlockFactoryInterface as TestSubject;
 
 /**
  * Tests {@see TestSubject}.
  *
- * @since 0.4
+ * @since 0.1
  */
-class StringBlockFactoryInterfaceTest extends TestCase
+class ContextAwareInterfaceTest extends TestCase
 {
     /**
      * Creates a new instance of the test subject.
      *
-     * @since 0.4
+     * @since 0.1
      *
      * @return TestSubject|MockObject
      */
@@ -31,15 +31,16 @@ class StringBlockFactoryInterfaceTest extends TestCase
     /**
      * Tests whether a valid instance of the test subject can be created.
      *
-     * @since 0.4
+     * @since 0.1
      */
     public function testCanBeCreated()
     {
         $subject = $this->createInstance();
 
         $this->assertInstanceOf(
-            TestSubject::class, $subject,
-            'A valid instance of the test subject could not be created.'
+            TestSubject::class,
+            $subject,
+            'Could not create a valid instance of the test subject.'
         );
     }
 }
