--- travis.php.ini
@@ -1,1 +1,0 @@
-memory_limit = 2048M
